LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-volume-styles-expected.txt

Testing the MacOSFullscreenMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(mediaControls.volumeSlider.element).left became "30px"
PASS window.getComputedStyle(mediaControls.volumeSlider.element).top is "19px"

PASS successfullyParsed is true

TEST COMPLETE
